feat: Add authCheckStatus function for checking authentication status
The `authCheckStatus` function is added to the `auth.ts` file in the `actions/auth` directory. This function makes a GET request to the `/auth/check-status` endpoint to check the authentication status. It returns the user token if the status is authenticated. This addition enhances the authentication functionality of the application.