feat: Update .gitignore and package.json
This commit updates the .gitignore file to include the .env and .yarn files. It also updates the package.json file to use the `expo run:android` and `expo run:ios` commands instead of `expo start --android` and `expo start --ios`. Additionally, it adds the `expo-camera` dependency to the project with version "~15.0.14" and updates the `react-native` dependency to version "0.74.5". These changes enhance the codebase by improving the project's build and development processes, as well as adding support for camera functionality.